account_service_console_application/requirements.txt
# Account Service Console Application Requirements

# Database
asyncpg>=0.28.0

# Core
python-dateutil>=2.8.2

# Encryption & Hashing
bcrypt>=4.0.0

# Terminal Colors
colorama>=0.4.6

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
